SENLA-17: перевел цену услуг на BigDecimal

+++ task-17/src/main/java/by/slava_borisov/hoteladmin/dao/AmenityDao.java
@@ -1,11 +1,13 @@
 package by.slava_borisov.hoteladmin.dao;
 
 import by.slava_borisov.hoteladmin.model.Amenity;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AmenityDao extends GenericDao<Amenity, Long> {
 
-    void updatePrice(Long amenityId, double newPrice);
+    void updatePrice(Long amenityId, BigDecimal newPrice);
 
     List<Amenity> findAllSortedByPrice();
