+++ task-17/src/main/java/by/slava_borisov/hoteladmin/dto/request/ChangePriceRequest.java
@@ -1,5 +1,7 @@
 package by.slava_borisov.hoteladmin.dto.request;
 
+import java.math.BigDecimal;
+
 public record ChangePriceRequest(
-        double newPrice
+        BigDecimal newPrice
 ) { }
