+++ task-17/src/main/java/by/slava_borisov/hoteladmin/service/AmenityService.java
@@ -4,6 +4,7 @@ import by.slava_borisov.hoteladmin.dto.AmenityDto;
 import by.slava_borisov.hoteladmin.dto.AmenityUsageDto;
 import by.slava_borisov.hoteladmin.util.SortCriteria;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -14,7 +15,7 @@ public interface AmenityService {
 
     List<AmenityDto> getAllAmenities();
 
-    void changeAmenityPrice(Long amenityId, double newPrice);
+    void changeAmenityPrice(Long amenityId, BigDecimal newPrice);
 
     Optional<AmenityDto> findAmenityById(Long amenityId);
