+++ task-17/src/main/java/by/slava_borisov/hoteladmin/service/RoomService.java
@@ -1,10 +1,11 @@
 package by.slava_borisov.hoteladmin.service;
 
-import by.slava_borisov.hoteladmin.dto.RoomDto;
 import by.slava_borisov.hoteladmin.dto.PriceDto;
+import by.slava_borisov.hoteladmin.dto.RoomDto;
 import by.slava_borisov.hoteladmin.model.RoomStatus;
 import by.slava_borisov.hoteladmin.util.SortCriteria;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -19,7 +20,7 @@ public interface RoomService {
 
     void setRoomStatus(Long roomId, RoomStatus status);
 
-    void changeRoomPrice(Long roomId, double newPrice);
+    void changeRoomPrice(Long roomId, BigDecimal newPrice);
 
     List<RoomDto> getAvailableRoomsOnDate(LocalDate date);
